requirements.txt
jsonrpclib
pycrypto
python-cjson
